requirements.txt
pandas
numpy
scikit-learn
joblib
flask
mlflow
black
pytest==8.3.2
PyYAML==6.0.1
setuptools==70.3.0
uvicorn==0.30.1
imbalanced-learn==0.12.3
fastapi==0.111.1